move files to correct folder
