+++ source/src/cip/cip_security/cipsecurity.c
@@ -36,7 +36,7 @@
 #include <stdio.h>
 
 #include "cipsecurity.h"
-#include "EtherNetIPSecurityObject/ethernetipsecurity.h"
+#include "ethernetipsecurity.h"
 
 #include "cipcommon.h"
 #include "endianconv.h"
